Restart triggers that fail
changelog_begin
changelog_end
